Split a wildcard implementation to xmvExpander
